Reduce the number DOM modifications
When using a crude string hash of data from the event database
we can reduce the number of DOM modifications.

The hash is stored in the relevant element and will be
compaired to the hash of a new query response from the
event database. If it is different, the resepsjon code
will update the elements it controls.